Add files via upload
This is a project I've been working on from the first day I learnt python. The python logic itself is pretty straight-forward, a bit bulky but understandable. It uses weighted averages to give you the probability of a driver achieving a podium in a given race.  Using some prompt engineering, I got the weighted averages from a model that Claude generated for me, using drivers' and constructors' data from kaggle.com, from 1950 to 2024. I've backtested it using historical results and it's accurate approximately 94% of the time. This is my current Magnum Opus.